Introduce CGFloat as a distinct struct type.
CGFloat is 32-bit on 32-bit architectures and 64-bit on 64-bit
architectures for historical reasons. Rather than having it alias
either Float (32-bit) or Double (64-bit), introduce a distinct struct
type for CGFloat. CGFloat provides a complete set of comparisons and
arithmetic operators (including tgmath functions), initializers allows
explicit conversion between it an Int, UInt, Float, and Double, as
well as conforming to all of the protocols that Float/Double do.

This formulation of CGFloat makes use of CGFloat
architecture-independent, although it still requires a number of casts.
Fixes <rdar://problem/17224725>

Swift SVN r19689

+++ stdlib/objc/CoreGraphics/CoreGraphics.swift
@@ -11,7 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 @exported import CoreGraphics
-
+import Darwin
 
 //===----------------------------------------------------------------------===//
 // CGGeometry
@@ -55,7 +55,7 @@ public extension CGVector : Equatable {
   static var zeroVector: CGVector { get { return CGVector(0, 0) } }
 
   init(_ dx: CGFloat, _ dy: CGFloat) {
-    self.init(dx: CGFloat(dx), dy: CGFloat(dy))
+    self.init(dx: dx, dy: dy)
   }
 
   init(_ dx: Int, _ dy: Int) {
